editoast: map: move cache key utils at crate root
Simplifies the map module.
